refactor(DataType): 컨트랙트에서 받아오는 data의 type을 모두 string으로 수정

+++ frontend/src/components/MyNFTPage/NFTCard.tsx
@@ -15,7 +15,7 @@ const NFTCard: FC<NFTCardProps> = ({ id, type, health, attack, price }) => {
 
   const onClickSell = async () => {
     if (!account) return;
-    if (!(parseInt(inputPrice) > 0)) {
+    if (!(parseFloat(inputPrice) > 0)) {
       alert('0 이상의 수를 입력해주세요');
       setInputPrice('');
       return false;
